% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegMars-package.R
\docType{package}
\name{ggsegMars-package}
\alias{ggsegMars}
\alias{ggsegMars-package}
\title{ggsegMars: MarsAtlas Cortical Parcellation for the 'ggseg' Ecosystem}
\description{
MarsAtlas cortical parcellation based on macroanatomical information with 41 regions per hemisphere for the 'ggseg' ecosystem. Provides a unified 'ggseg_atlas' object with both 2D polygon geometry and 3D vertex indices, for use with 'ggseg' and 'ggseg3d'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggsegverse/ggsegMars}
  \item Report bugs at \url{https://github.com/ggsegverse/ggsegMars/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

}
\keyword{internal}
