% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{marsatlas_cerebellar}
\alias{marsatlas_cerebellar}
\title{MarsAtlas Cerebellar Parcellation}
\usage{
marsatlas_cerebellar()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (subcortical).
}
\description{
Cerebellar regions from the MarsAtlas parcellation.
Contains 2D polygon geometry and 3D meshes.
}
\examples{
marsatlas_cerebellar()
}
\references{
Auzias G, Coulon O, Brovelli A (2016).
MarsAtlas: A cortical parcellation atlas for functional mapping.
\doi{10.1002/hbm.23121}
}
\seealso{
Other ggseg_atlases: 
\code{\link{marsatlas_cortical}()},
\code{\link{marsatlas_subcortical}()}
}
\concept{cerebellar_atlases}
\concept{ggseg_atlases}
