% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{marsatlas_cortical}
\alias{marsatlas_cortical}
\title{MarsAtlas Cortical Parcellation}
\usage{
marsatlas_cortical()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\description{
Cortical parcellation based on macroanatomical information from
Auzias et al. (2016), with 41 regions per hemisphere.
Contains 2D polygon geometry for \code{\link[ggseg:ggbrain]{ggseg::geom_brain()}} and
3D vertex indices for \code{\link[ggseg3d:ggseg3d]{ggseg3d::ggseg3d()}}.
}
\examples{
marsatlas_cortical()
}
\references{
Auzias G, Coulon O, Brovelli A (2016).
MarsAtlas: A cortical parcellation atlas for functional mapping.
\doi{10.1002/hbm.23121}
}
\seealso{
Other ggseg_atlases: 
\code{\link{marsatlas_subcortical}()}
}
\concept{ggseg_atlases}
