% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{marsatlas_subcortical}
\alias{marsatlas_subcortical}
\title{MarsAtlas Subcortical Structures}
\usage{
marsatlas_subcortical()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (subcortical).
}
\description{
Subcortical structures from the MarsAtlas parcellation.
Contains 2D polygon geometry and 3D meshes.
}
\examples{
marsatlas_subcortical()
}
\references{
Auzias G, Coulon O, Brovelli A (2016).
MarsAtlas: A cortical parcellation atlas for functional mapping.
\doi{10.1002/hbm.23121}
}
\seealso{
Other ggseg_atlases: 
\code{\link{marsatlas_cortical}()}
}
\concept{ggseg_atlases}
